broke down the main file into 3 different files moved classes to classes.py moved functions to functions.py the main function sits on main.py
